feat: Migrate backend to Vercel Serverless Functions
This definitive commit refactors the application's backend from a stateful Express server to a modern, robust Vercel serverless architecture. This resolves all previously identified deployment and runtime crashes.

Key Changes:
- The entire backend is now composed of individual serverless functions located in the `/api` directory (`generate.ts`, `edit.ts`, `segment.ts`).
- Shared logic for API key handling and prompt building has been extracted into a reusable helper file at `/api/_lib/helpers.ts`.
- A `vercel.json` file has been added to configure the necessary rewrites for the API routes.
- The old, problematic server infrastructure (`server.ts`, `tsconfig.server.json`, `concurrently`, `express`) has been removed.
- `package.json` has been updated to use `vercel dev` for local development, which correctly simulates the Vercel environment.
- The `vite.config.ts` has been cleaned up by removing the now-redundant proxy configuration.
