+++ package.json
@@ -4,13 +4,8 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "dev": "concurrently \"npm:dev:frontend\" \"npm:dev:backend\"",
-    "dev:frontend": "vite",
-    "dev:backend": "node --loader ts-node/esm server.ts",
-    "build": "npm run build:frontend && npm run build:backend",
-    "build:frontend": "vite build",
-    "build:backend": "tsc --project tsconfig.server.json && echo '{\"type\": \"module\"}' > dist-server/package.json",
-    "start": "node dist-server/server.js",
+    "dev": "vercel dev",
+    "build": "vite build",
     "lint": "eslint .",
     "preview": "vite preview"
   },
@@ -23,7 +18,6 @@
     "@tanstack/react-query": "^5.85.5",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "express": "^5.1.0",
     "fabric": "^6.7.1",
     "idb-keyval": "^6.2.2",
     "konva": "^9.3.22",
@@ -37,21 +31,19 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.9.1",
-    "@types/express": "^5.0.3",
     "@types/react": "^18.3.5",
     "@types/react-dom": "^18.3.0",
     "@vitejs/plugin-react": "^4.3.1",
     "autoprefixer": "^10.4.18",
-    "concurrently": "^9.2.1",
     "eslint": "^9.9.1",
     "eslint-plugin-react-hooks": "^5.1.0-rc.0",
     "eslint-plugin-react-refresh": "^0.4.11",
     "globals": "^15.9.0",
     "postcss": "^8.4.35",
     "tailwindcss": "^3.4.1",
-    "ts-node": "^10.9.2",
     "typescript": "^5.5.3",
     "typescript-eslint": "^8.3.0",
+    "vercel": "^48.1.6",
     "vite": "^5.4.2"
   }
 }
