--- tsconfig.server.json
@@ -1,13 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "module": "NodeNext",
-    "moduleResolution": "NodeNext",
-    "outDir": "dist-server",
-    "noEmit": false,
-    "isolatedModules": false,
-    "resolveJsonModule": true
-  },
-  "include": ["server.ts"],
-  "exclude": ["src", "node_modules"]
-}
